refactors containers dir
