+++ opentrons_sdk/containers/__init__.py
@@ -1,0 +1,6 @@
+from opentrons_sdk.protocol.robot import Robot
+
+
+def load(container_name, slot):
+    protocol = Robot.get_instance()
+    return protocol.add_container(slot, container_name)
