+++ opentrons_sdk/protocol/robot.py
@@ -1,7 +1,7 @@
 import copy
 
-from opentrons_sdk.containers import legacy_containers, placeable
 import opentrons_sdk.drivers.motor as motor_drivers
+from opentrons_sdk.containers import legacy_containers, placeable
 from opentrons_sdk.protocol.command import Command
 from opentrons_sdk.util import log
