+++ tests/opentrons_sdk/integration/test_protocol.py
@@ -1,8 +1,9 @@
 import unittest
+from opentrons_sdk import containers
 
 from opentrons_sdk.protocol import Robot
 from opentrons_sdk.containers.placeable import Container, Deck
-from opentrons_sdk.labware import containers, instruments
+from opentrons_sdk.labware import instruments
 
 
 class ProtocolTestCase(unittest.TestCase):
