+++ tests/opentrons_sdk/labware/test_instruments.py
@@ -1,8 +1,9 @@
 import unittest
 from unittest import mock
 from unittest.mock import call
+from opentrons_sdk import containers
 
-from opentrons_sdk.labware import containers, instruments
+from opentrons_sdk.labware import instruments
 from opentrons_sdk.protocol import Robot
 from opentrons_sdk.protocol.command import Command
